Changed when events are fired.
Events are now fired from lastTime (exclusive) to time (inclusive). This allows an event at the end of an animation to be fired before the next animation starts. The "complete" callback will happen before the event though, because "complete" happens in update() and events are triggered in apply().

Note events are not fired for the previous animation during mixing (crossfading).