Changed Batch to SpriteBatch
I am not sure the issue here and I apologize that I am fairly new to Github. But when trying to compile the Spine libGDX runtime last night I noticed this error.

Very simple fix just change Batch to SpriteBatch at lines 41 and 85

I think the new libGDX gets rid of the having a public Batch class so you use SpriteBatch. If this is not the case, please let me know! Thanks :)