Submesh triangles may not be set when not rendering Meshes.
Hi Nathan!
One comment to this return: Is it possible that it should be inside the inner "if" block? We have one issue with a character which doesn't use mesh attachments, but shows some strange behaviour when I turn off the renderMeshes option. When I move the return in the inner if block the issue is gone which makes sense because otherwise the triangles for the sub mesh may never be set if the check "submesh.firstVertex != firstVertex || submesh.triangleCount < triangleCount" fails.

Cheers
Christian