Schedule update calls when entering the scene.
If a node is removed from the scene, then Cocos2d-x will unschedule its
selectors. In order to have the animation still playing if the node is
inserted again in the scene, the call to scheduleUpdate() in SkeletonRenderer
must then be done in SkeletonRenderer::onEnter() instead of at initialization
time.

This commit applies this change, making the following code to work:

    void testAnimatedWhenReinserted( SkeletonRenderer& skeleton )
    {
        skeleton.retain();
        cocos2d::CCNode* const parent{ skeleton.getParent() };
        skeleton.removeFromParent();
        parent->addChild( &skeleton );
        skeleton.release();
    }