Updated AttachmentTimeline to check whether expected attachment is set.
Until now it was discarding the set of the attachment if lastTime is after keyTime. I'm having issues with this behavior when the skeleton is set to pose manually. After that, the attachment is the one used in setup intead of the expected one until the next keyframe.

For more information on the scenario. I'm using this to revert to pose only specific tracks. Doing the following.
1 -  Clear the Track To Revert
2 -  Revert Skeleton to pose
3 -  Update(0) this way the rest of the tracks apply to the skeleton again

I've made several characterization tests of this timeline before changing it, I can pass them to you if you want.