Apply remaining time to next animation.
Previously one frame worth of delta time was lost.

When the end of an animation is passed, one frame goes by so the animation has a chance to apply its final pose and fire any events on its last few frames. The second animation starts using the amount of time since the actual end of the first animation.

#154