[Unity] lastMaterial != null hotspot.
I was profiling a variant of this SkeletonRenderer code and this seemed to be a hotspot.
`UnityEngine.Object != null` is overloaded behind the scenes and has a pretty deep call stack and does a bunch of things. This scales badly as the number of attachments increases. See this UnityEngine code for the details: https://github.com/MattRix/UnityDecompiled/blob/master/UnityEngine/UnityEngine/Object.cs#L180

`vertexCount > 0` effectively serves the same purpose (gate the operation until the first renderable attachment gets processed).