[unity] Unity Sprite slices for atlas textures.
The new "Apply Regions as Texture Sprite Slices" button adds Sprite slices to all atlas page Textures of an AtlasAsset in Unity Editor.
Clicking the button updates existing sprites if it detects name-matches. Adds ones that don't exist. It does not remove non-matches.

Sprite slices are not auto-applied on atlas import. Users have to opt-in by clicking the button.
This is also to communicate that spine-unity does not currently rely on the Unity sprite system. Unity 5.5 has promising API changes to this end but this cannot be a standard spine-unity feature yet.

This replaces the functionality of the old Region Baking as prefabs with mesh assets. Original functionality has been precompiler-defined out. Region baking code will be removed in a future version.