[unity] Add option to not pause spine with timeline
There is already an option with hold the animation, but there is no
option to not pause the animation while the Timeline director is paused.

An example use-case where this option is important is when the timeline
is paused while waiting for an user input, but there is still a need to
play an idle loop animation on a spine skeleton.