[unity] Initial support for prefab preview thumbnails
In order to support prefab thumbnails MeshFilter.sharedMesh needs to be
saved as an asset, otherwise the scene mesh will be used by unity, which
creates a type mismatch. To overcome this problem a prefab asset post
processor is introduced to generate and save a new mesh with the prefab.

When a prefab is created/updated this post process script is run and
generates a new skeleton mesh and saves the one stored in MeshFilter
component for each game object with SkeletonRenderer component attached.
This way unity can use this saved in prefab mesh to generate preview
thumbnail for the prefab.