[unity] Implement build preprocessor for prefabs cleanup
With the new prefab processing there is an issue where the mesh tries to
get into the build. To prevent this a build pre/post process is
introduced that will clean up the prefab from the meshes before the
build and restore the prefab meshes after the build.

This change reverts bf70a62f1 and is related to #1273,#1931.