Removed spine-lua, spine-solar2d, spine-love, spine-as3, and spine-starling.
These were deprecated in Spine 4.0 and are removed as of 4.1. This is because use of these runtimes has been declining over the years due to discontinuation of the products they are built upon, like Corona and Flash. We want to free up development resources to work on new features and new runtimes.