feat (loadTextureAtlas) add option to use file alias (#2019)
For Construct 3, in preview mode, the texture file names are changed to blobs with completely different filenames. So, loadTextureAtalas will not work because the file names don't match the page.name.  When Construct 3 is not in preview mode, the normal file names will match the page.name.

To get around this issue, added an optional parameter to loadTextureAtlas, fileAlias, this will allow the Construct 3 plug-in to create a fileAlias object which maps between page.name and the C3 project file names, whether they are aliased or not. If this is not needed, the function performs as it did before.

Co-authored-by: MikalDev <keg@kindeyegames.com>