[ts] Use existing, explicit atlas data instead of relying on an image's current width/height
This fixes a bug where UVs would sometimes be wrong if an image's size is not the same as when originally exported.  This could happen in the wild to _any_ project, due to mobile ISPs/browsers that proxy requests and resize images before sending them to clients.
Fixing this also allows `spine-core` to work with engines that:
* use embedded, low-res textures while waiting for high res textures to load
* automatically provide lower resolution textures to low-end devices or low-resolution displays