Add WasNextApplied to TrackEntry
This allows users to check when/if the next TrackEntry in a sequence of TrackEntries is applied (i.e. TrackEntries added to the currently playing TrackEntry using AnimationState.AddAnimation.

This is intended to avoid exposing any additional internal members that could be confusing to some users.

This is to avoid the issue that when TrackEntry has a non-null next member, using TrackEntry.Next.WasApplied will always return true, regardless of whether it has ever actually been applied.