Remove SkeletonSprite from the Juggler on dispose
When we dispose of the SkeletonSprite, we should remove it from the juggler - since this is the default thing that would be calling advanceTime().  If the SkeletonSprite is not in the juggler, nothing happens, but if we don't remove it - we now have a memory leak (and we are calling advanceTime on a item that it disposed).

You could make an argument then when we create a SkeletonSprite, it should be added to the Starling juggler automatically.  However, I can see arguments to NOT do this (i.e. you aren't using the default juggler, or you are using something else to call advanceTime).